requirement.txt
streamlit
google-generativeai
pypdf2
python-dotenv